Effect fixes for 450-LED strips, GPS-driven shows, and WiFi power policy
Effect position maths used uint8_t throughout, so on 450-LED strips the
meteor shower, rainbow comet, matrix rain, lava lamp, kaleidoscope,
pulse wave and colour explosion all wrapped positions past 255 back onto
the front of the strip. Positions are wide types now. Along the way:
colour explosion re-triggers when the wave clears the strip (the old
`% 500 == 0` check almost never fired), comet trails fade to black
instead of wrapping bright, the palette/spectrum stream hue advances
once per frame so an 8-strip rig cycles colours at the same rate as a
single-strip build, and apply_scene_updates writes palette_stream's own
duration instead of leaning on union aliasing.

GPS modes now actually follow the GPS: loop() re-evaluates them every
second - speedometer re-blends its colour, position_status and the GPS
speed animation adjust pace via setSpeed(), which changes the frame
period in place so the animation never visibly restarts. The basic
status chunk reports gpsLightSpdEn/gpsLowSpd/gpsTopSpd so the app can
learn them, and a 2-byte palette write is only a numeric id when the
byte is in id range - the palette literally named "r" is selectable.

Battery: wearables check for updates at boot then turn the WiFi radio
fully off (staying associated next to BLE was the single biggest drain);
the app's check-now or fresh credentials bring it up and linger five
minutes so espota still works on demand. Mains-powered signs keep the
always-on behaviour. Failed connects also power the radio down rather
than leaving auto-reconnect hunting. CPU clock drops to 160 MHz
(-DBM_CPU_MHZ to override) and the main loop yields between frames -
effects gate on millis(), so pacing adds at most 1 ms of frame jitter.

Encoder brightness respects the stored max-brightness instead of a
hard-coded 125 cap, the OTA red screen is one showColor at a relaxed
pace instead of a per-target fill_solid ladder, and the C6 boot log
reports the real LED count.

+++ libraries/BurningManLEDs/src/LightShow.h
@@ -368,11 +368,13 @@ private:
     size_t heat_array_size_;       // Size of heat array
     uint8_t *meteor_positions_;    // Positions of meteors
     uint8_t *meteor_trails_;       // Trail intensities
-    uint8_t pulse_center_;         // Center position for pulse waves
+    // LED positions are uint16_t: strips run to 450 LEDs, and a uint8_t here
+    // wrapped every position past 255 back onto the front of the strip.
+    uint16_t pulse_center_;        // Center position for pulse waves
     uint8_t matrix_drops_[64];     // Matrix rain drop positions (max 64 drops)
     uint8_t plasma_offset_;        // Offset for plasma clouds
     float noise_x_, noise_y_;      // For smooth noise effects
-    uint8_t explosion_center_;     // Center of color explosion
+    uint16_t explosion_center_;    // Center of color explosion
     uint8_t spiral_angle_;         // Current spiral rotation
 };
